refactor: Move mocks to their respective packages
This commit moves the mock structures and methods from the test package
to their respective packages. This change makes the codebase more
organized and easier to navigate. The MockEntryCrypto was moved to the
services package and the MockEntryModel was moved to the models package.
