+++ internal/models/mock.go
@@ -1,27 +1,12 @@
-package test
+package models
 
 import (
 	"context"
 	"database/sql"
 	"time"
 
-	"github.com/Ajnasz/sekret.link/internal/models"
 	"github.com/stretchr/testify/mock"
 )
-
-type MockEntryCrypto struct {
-	mock.Mock
-}
-
-func (m *MockEntryCrypto) Encrypt(data []byte) ([]byte, error) {
-	args := m.Called(data)
-	return args.Get(0).([]byte), args.Error(1)
-}
-
-func (m *MockEntryCrypto) Decrypt(data []byte) ([]byte, error) {
-	args := m.Called(data)
-	return args.Get(0).([]byte), args.Error(1)
-}
 
 type MockEntryModel struct {
 	mock.Mock
@@ -33,14 +18,14 @@ func (m *MockEntryModel) CreateEntry(
 	UUID string,
 	data []byte,
 	remainingReads int,
-	expire time.Duration) (*models.EntryMeta, error) {
+	expire time.Duration) (*EntryMeta, error) {
 	args := m.Called(ctx, tx, UUID, data, remainingReads, expire)
-	return args.Get(0).(*models.EntryMeta), args.Error(1)
+	return args.Get(0).(*EntryMeta), args.Error(1)
 }
 
-func (m *MockEntryModel) ReadEntry(ctx context.Context, tx *sql.Tx, UUID string) (*models.Entry, error) {
+func (m *MockEntryModel) ReadEntry(ctx context.Context, tx *sql.Tx, UUID string) (*Entry, error) {
 	args := m.Called(ctx, tx, UUID)
-	return args.Get(0).(*models.Entry), args.Error(1)
+	return args.Get(0).(*Entry), args.Error(1)
 }
 
 func (m *MockEntryModel) UpdateAccessed(ctx context.Context, tx *sql.Tx, UUID string) error {
